requirements.txt
discord.py==2.3.2
dnspython==2.8.0
python-dotenv==1.0.1
aiohttp==3.9.1
psutil==5.9.8
python-whois==0.8.0
gputil==1.4.0
setuptools
fastapi
uvicorn
geoip2
yt-dlp
PyNaCl